# [2.0] Improve Json Format new registry performance a bit