@frankmayer 2017-07-09
The changes look fine to me. Nice little improvement. :)